+++ Posh-IBWAPI/Public/Set-IBObject.ps1
@@ -14,8 +14,8 @@ function Set-IBObject
 
         [Alias('fields')]
         [string[]]$ReturnFields,
-        [Alias('base')]
-        [switch]$ReturnBaseFields,
+        [Alias('base','ReturnBaseFields')]
+        [switch]$ReturnBase,
         [switch]$BatchMode,
         [ValidateRange(1,2147483647)]
         [int]$BatchGroupSize = 1000,
@@ -39,14 +39,14 @@ function Set-IBObject
 
         # process the return fields
         if ($ReturnFields.Count -gt 0) {
-            if ($ReturnBaseFields) {
+            if ($ReturnBase) {
                 $querystring = "?_return_fields%2B=$($ReturnFields -join ',')"
             }
             else {
                 $querystring = "?_return_fields=$($ReturnFields -join ',')"
             }
         }
-        elseif ($ReturnBaseFields) {
+        elseif ($ReturnBase) {
             $querystring = "?_return_fields%2B"
         }
 
@@ -96,7 +96,7 @@ function Set-IBObject
         # build the 'args' value for each object
         $retArgs = @{}
         if ($ReturnFields.Count -gt 0) {
-            if ($ReturnBaseFields) {
+            if ($ReturnBase) {
                 $retArgs.'_return_fields+' = $ReturnFields -join ','
             } else {
                 $retArgs.'_return_fields'  = $ReturnFields -join ','